Karen AI-Voice Assistant/requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
pyjokes
opencv-python
requests
pyautogui
sounddevice
edge-tts
pygame
google-genai
psutil
speedtest-cli
Pillow
gnews
yt-dlp
pypdf
